PCI: cadence: Correct probe behaviour when failing to get PHY
Test the correct value to see whether the PHY get operation failed.

Use devm_phy_get() instead of devm_phy_optional_get(), since it is only
called if the PHY name is given in the devicetree and so should exist.

If failure when getting or linking PHY, put any PHYs which were
already got and unlink them.

Fixes: dfb80534692ddc5b ("PCI: cadence: Add generic PHY support to host and EP drivers")
Signed-off-by: Alan Douglas <adouglas@cadence.com>
[lorenzo.pieralisi@arm.com: updated commit log]
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>