PCI/ERR: Always use the first downstream port
The link reset always used the first bridge device, but AER broadcast error
handling may have reported an endpoint.  This means the reset may hit
devices that were never notified of the impending error recovery.

Use the first downstream port in the hierarchy considered reliable.  An
error detected by a switch upstream port should mean it occurred on its
upstream link, so the patch selects the parent device if the error is not a
root or downstream port.

This allows two other clean-ups.  First, error handling can only run on
bridges so this patch removes checks for endpoints.  Second, the first
accessible port does not inherit the channel error state since we can
access it, so the special cases for error detect and resume are no longer
necessary.

Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Sinan Kaya <okaya@kernel.org>