nfit/libnvdimm: add set passphrase support for Intel nvdimms
Add support for setting and/or updating passphrase on the Intel nvdimms.
The passphrase is pulled from userspace through the kernel key
management.  We trigger the update via writing "update <old_keyid>
<new_keyid>" to the sysfs attribute "security". If no <old_keyid> exists
(for enabling security) then a 0 should be used. The state of the
security can also be read via the "security" attribute. libnvdimm will
generically support the key_change API call.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>