parisc: Add alternative coding when running UP
This patch adds the necessary code to patch a running SMP kernel
at runtime to improve performance when running on a single CPU.

The current implementation offers two patching variants:
- Unwanted assembler statements like locking functions are overwritten
  with NOPs. When multiple instructions shall be skipped, one branch
  instruction is used instead of multiple nop instructions.

- Some pdtlb and pitlb instructions are patched to become pdtlb,l and
  pitlb,l which only flushes the CPU-local tlb entries instead of
  broadcasting the flush to other CPUs in the system and thus may
  improve performance.

Live-patching is done early in the boot process, just after having run
the system inventory. No drivers are running and thus no external
interrupts should arrive. So the hope is that no TLB exceptions will
occur during the patching. If this turns out to be wrong we will
probably need to do the patching in real-mode.

Signed-off-by: Helge Deller <deller@gmx.de>