cifs: check how much EA space we already use before adding more EAs
When setting/adding an EA, check how much EA space we arelady have on this
file and make sure that if we add this new EA that the result will still
fit inside a SMB2_MAX_EA_BUF.

In the client we limit the maximum buffer to currently 64k.
When we store these on the server we have no way of knowing if/what the
servers limit for EAs are. They could be, and often are, much larger than
64k.
This change is to return an error and not set the EA if we suspect that
the result could overflow the buffer.

This prevents the very annoying behaviour that we can add EAs well
beyond the maximum clientside buffer and thus, at that stage make
it is no longer query or list the existing EAs any more.

The check for overflow is overly conservative :
        if (len + rsp_iov[0].iov_len > SMB2_MAX_EA_BUF) {
since the actual size of the EAs are a bit less than iov_len
but this is simpler to read.

The only effect of this is that we will stop allowing adding new EAs ~100
bytes earlier than stricly needed which does not matter much for 64k
buffers.

Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Steve French <stfrench@microsoft.com>